Mock import of `soundfile'.
Set autodoc to mock the import of `soundfile'. This is necessary,
because `cffi', which is a dependecy of `soundfile', raises an OSError
if it cannot find the underlying C library `sndfile'. This typically the
case on RTD.

Additionally, reset the project name to all lower-case and set typehints
as signatures.